Crate polkadot_collator_protocol

The Collator Protocol allows collators and validators talk to each other. This subsystem implements both sides of the collator protocol.

Structs

CollatorEvictionPolicy

A collator eviction policy - how fast to evict collators which are inactive.

CollatorProtocolSubsystem

The collator protocol subsystem.

Enums

ProtocolSide

What side of the collator protocol is being engaged